+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/apoorv/todo-frontend/src/index.js":"1"},{"size":19711,"mtime":1607870091078,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1mr5qeb","/home/apoorv/todo-frontend/src/index.js",[]]
+[{"/home/apoorv/todo-frontend/src/index.js":"1","/home/apoorv/todo-frontend/src/components/Header.js":"2","/home/apoorv/todo-frontend/src/components/Home.js":"3","/home/apoorv/todo-frontend/src/components/SignUp.js":"4","/home/apoorv/todo-frontend/src/App.js":"5","/home/apoorv/todo-frontend/src/components/TodoList/TodoList.js":"6","/home/apoorv/todo-frontend/src/components/TodoList/AddTodo/AddTodo.js":"7","/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItem/TodoItem.js":"8","/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItems.js":"9","/home/apoorv/todo-frontend/src/components/TodoList/Footer/Footer.js":"10"},{"size":254,"mtime":1607881922640,"results":"11","hashOfConfig":"12"},{"size":427,"mtime":1607881364155,"results":"13","hashOfConfig":"12"},{"size":3244,"mtime":1607883009634,"results":"14","hashOfConfig":"12"},{"size":4274,"mtime":1607881355699,"results":"15","hashOfConfig":"12"},{"size":1419,"mtime":1607888874151,"results":"16","hashOfConfig":"12"},{"size":6213,"mtime":1607883240291,"results":"17","hashOfConfig":"12"},{"size":2126,"mtime":1607883132490,"results":"18","hashOfConfig":"12"},{"size":908,"mtime":1607889123446,"results":"19","hashOfConfig":"12"},{"size":491,"mtime":1607889049462,"results":"20","hashOfConfig":"12"},{"size":1107,"mtime":1607881366111,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1mr5qeb",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/apoorv/todo-frontend/src/index.js",[],"/home/apoorv/todo-frontend/src/components/Header.js",[],"/home/apoorv/todo-frontend/src/components/Home.js",[],"/home/apoorv/todo-frontend/src/components/SignUp.js",[],"/home/apoorv/todo-frontend/src/App.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/TodoList.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/AddTodo/AddTodo.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItem/TodoItem.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItems.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/Footer/Footer.js",[]]
